change long to int in responcedto
change long to int in responcedto

+++ WebAPI/Models/ResponseDTO.cs
@@ -3,10 +3,10 @@ namespace DNDProject.Models
 {
     public class ResponseDTO
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Answer { get; set; } = string.Empty;
-        public long QuestionId { get; set; }
-        public long SurveyId { get; set; }  
+        public int QuestionId { get; set; }
+        public int SurveyId { get; set; }  
         public DateTime ResponseDate { get; set; }  
     }
 }
